face-recognition/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
distlib==0.3.4
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.4.2
idna==3.3
joblib==1.1.0
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.1
pyparsing==3.0.6
redis==4.1.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
virtualenv==20.13.0
wrapt==1.13.3
